pyproject.toml
[project]
name = "hybrid-calorie-app"
version = "0.1.0"
description = "Hybrid calorie tracking (LLM + structured DB)"
readme = "architecture.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"
markers = [
    "integration: live HTTP (requires API keys)",
    "live_usda: real USDA FDC HTTP (skip autouse OFF/USDA stub; needs USDA_FDC_API_KEY)",
]
